取得賽事資料api url調整

+++ database/seeders/EventSeeder.php
@@ -24,55 +24,61 @@ class EventSeeder extends Seeder
         try {
             DB::statement('SET FOREIGN_KEY_CHECKS=0;');
     
-            $response = Http::get('http://dc942e3569de.ngrok.io/api/events');
-    
-            $resdatas = $response->json();
-    
-            $eventsColumns = DB::getSchemaBuilder()->getColumnListing('events');
-            $eventsDistancesColumns = DB::getSchemaBuilder()->getColumnListing('events_distances');
-    
-            if (count($resdatas) > 0) {
-                app(Event::class)->truncate();
-                app(EventDistance::class)->truncate();    
-    
-                foreach($resdatas as $data) {
-                    if (is_array($data)) {
-                        $id = uniqid();
-                        $formData = [
-                            'id' => $id,
-                        ];
-                        $distanceFormData = [];    
-                        foreach($data as $key => $value) {
-                            if(in_array($key, $eventsColumns)) {
-                                $formData[$key] = $value;
-                            }
-                            if ($key === 'distances') {
-                                foreach($data[$key] as $distance) {
-                                    $distanceData = [
-                                        'id' => uniqid(),
-                                        'event_id' => $id,
-                                    ];
-                                    foreach($distance as $key => $value) {
-                                        if(in_array($key, $eventsDistancesColumns)) {
-                                            $distanceData[$key] = $value;
+            $response = Http::get(env('NODE_URL'));
+
+            if ($response->status() === 200) {
+
+                $resdatas = $response->json();
+        
+                $eventsColumns = DB::getSchemaBuilder()->getColumnListing('events');
+                $eventsDistancesColumns = DB::getSchemaBuilder()->getColumnListing('events_distances');
+        
+                if (count($resdatas) > 0) {
+                    app(Event::class)->truncate();
+                    app(EventDistance::class)->truncate();    
+        
+                    foreach($resdatas as $data) {
+                        if (is_array($data)) {
+                            $id = uniqid();
+                            $formData = [
+                                'id' => $id,
+                            ];
+                            $distanceFormData = [];    
+                            foreach($data as $key => $value) {
+                                if(in_array($key, $eventsColumns)) {
+                                    $formData[$key] = $value;
+                                }
+                                if ($key === 'distances') {
+                                    foreach($data[$key] as $distance) {
+                                        $distanceData = [
+                                            'id' => uniqid(),
+                                            'event_id' => $id,
+                                        ];
+                                        foreach($distance as $key => $value) {
+                                            if(in_array($key, $eventsDistancesColumns)) {
+                                                $distanceData[$key] = $value;
+                                            }
                                         }
+                                        array_push($distanceFormData, $distanceData);
                                     }
-                                    array_push($distanceFormData, $distanceData);
+                                }
+                            }
+                            app(Event::class)->create($formData);
+                            if (count($distanceFormData) > 0) {
+                                foreach($distanceFormData as $data) {
+                                    app(EventDistance::class)->create($data);
                                 }
                             }
                         }
-                        app(Event::class)->create($formData);
-                        if (count($distanceFormData) > 0) {
-                            foreach($distanceFormData as $data) {
-                                app(EventDistance::class)->create($data);
-                            }
-                        }
                     }
+                    Log::info('賽事資料更新完成');
+                } else {
+                    Log::info('無法取得賽事資料');
                 }
-                Log::info('賽事資料更新完成');
             } else {
-                Log::info('無法取得賽事資料');
+                Log::info('無法取得賽事資料:無法連線');
             }
+    
     
             DB::statement('SET FOREIGN_KEY_CHECKS=1;');
         } catch (Throwable $e) {
